Refactor the components & the test suites accordingly
- I've refactored the front-end from a single `Reviews` components
to include a `Review` subcomponent with its own internal state
improving the structure & separating the concerns of the components.

- I've added a new feature to the like button which is if the user
clicks the like button in increases the `likes_count`
& turns to a "unlike" button so that if it's clicked again it will turn
back to a "like" button & decrease the `likes_count` again.

- I've updated the server code to account for this newly added feature.

- I've updated the server suite & the front-end suite to comply with
the newly refactoed components.
